Create setJSLink.js
Creating a new site column with the JSLink property set is easy enough with server-side code or even PowerShell but if you have no server access then this script can be simply pasted into the console through dev tools of the browser.

Set the name of the site column and the path to the script file.  You will need to be logged in as a site collection admin and may also need to be on a list page to ensure you get the correct ctx.